+++ Changelog.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Updated credits
 - Consistently use `GlStateManager`
 - Snoozer file now uses `NAME` instead of `VERSION`
+- Optimized check for the warning screen
 
 ### Fixed